+++ packages/expo/src/hooks/useAuth.ts
@@ -1,4 +1,4 @@
-import { useAuth as useAuthBase } from '@clerk/clerk-react';
+import { useAuth as useAuthBase } from '@clerk/react';
 import { isNetworkError } from '@clerk/shared/error';
 import type { GetToken, GetTokenOptions, UseAuthReturn } from '@clerk/types';