+++ packages/nextjs/src/types.ts
@@ -1,4 +1,4 @@
-import type { ClerkProviderProps } from '@clerk/clerk-react';
+import type { ClerkProviderProps } from '@clerk/react';
 import type { Without } from '@clerk/types';
 
 export type NextClerkProviderProps = Without<ClerkProviderProps, 'publishableKey'> & {