+++ scripts/snapshot.mjs
@@ -14,10 +14,10 @@ const snapshot = `---
 '@clerk/nextjs': patch
 '@clerk/shared': patch
 '@clerk/themes': patch
-'@clerk/clerk-react': patch
+'@clerk/react': patch
 '@clerk/remix': patch
 '@clerk/types': patch
-'@clerk/clerk-expo': patch
+'@clerk/expo': patch
 '@clerk/express': patch
 '@clerk/testing': patch
 '@clerk/elements': patch